requirements.txt
Django==4.2
django-environ==0.10.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
gunicorn==21.2.0
Pillow==9.3.0
psycopg2==2.9.7
drf-nested-routers==0.93.4